+++ NEWS.md
@@ -11,6 +11,7 @@ None
 5. Fix #2 - Add function `tidy_autoplot()`
 6. Fix #11 - Add function `tidy_distribution_summary_tbl()`
 7. Fix #10 - Add function `tidy_empirical()`
+8. Fix #13 - Add function `tidy_uniform()`
 
 * Added a `NEWS.md` file to track changes to the package.